% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compak-package.R
\docType{package}
\name{compak-package}
\alias{compak-package}
\alias{compak}
\title{Conway-Maxwell-Poisson Associated Kernel}
\description{
Conway-Maxwell-Poisson Associated Kernel
}
\references{
Huang, A., Sippel, L. and Fung, T. (2020). \code{compak}:
Conway-Maxwell-Poisson Associated Kernel. R package version 0.1

Huang, A., Sippel, L. and Fung, T. (2020). A consistent second-order discrete kernel smoother.
Under revision.
}
\keyword{package}
