% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compak_dcomp.R
\name{compak_dcomp}
\alias{compak_dcomp}
\title{Density of Conway-Maxwell-Poisson distribution with parameter \code{mu} and \code{nu} with a
\code{compak} setting}
\usage{
compak_dcomp(x, mu, nu)
}
\arguments{
\item{x}{vector of quantiles}

\item{mu, nu}{numeric scalars: mean and dispersion parameters.}
}
\value{
The density of the cmp distribution.
}
\description{
Density of Conway-Maxwell-Poisson distribution with parameter \code{mu} and \code{nu} with a
\code{compak} setting
}
\keyword{internal}
