% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compak_evalkernel.R
\name{compak_evalkernel}
\alias{compak_evalkernel}
\title{Calculates the kernel at a key point}
\usage{
compak_evalkernel(key, counts, x, nu)
}
\arguments{
\item{key}{character: where the kernel should be centred.}

\item{counts}{list of distinct count values; names are the observed values; elements are the corresponding frequencies}

\item{x}{numeric vector :the points of the grid at which the density is to be estimated.}

\item{nu}{numeric: CMP_mu kernel has dispersion \code{nu}}
}
\value{
The kernel values.
}
\description{
This function evaluates the kernel, which is centred at \code{key}, over a grid of points using the counts provided in a list.
}
\keyword{internal}
