% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compak_evalpmf.R
\name{compak_evalpmf}
\alias{compak_evalpmf}
\title{Fits CMP_mu discrete kernel smoother for a sample of counts}
\usage{
compak_evalpmf(a.sample, x = NULL, h = NULL, nu = NULL, parallel = FALSE)
}
\arguments{
\item{a.sample}{numeric vector: the data sample from which estimate is to be computed.}

\item{x}{Either \code{NULL} or a numeric vector: the points of the grid at which the density is to be estimated. If none is provided, the range of \code{a.sample} will be used.}

\item{h, nu}{numeric: the bandwidth or smoothing parameter. Only one is needed and they are related by nu = 1/h.}

\item{parallel}{logical; determines whether to use parallel process to evaluate the kde}
}
\value{
Return the estimated p.m.f. values.
}
\description{
This function computes the density using the CMP_mu discrete kernel smooth over a grid of points using the given bandwidth \code{h} or dispersion parameter \code{nu}.
}
\examples{
data(days)
# The bandwidth h can be the one obtained by cross validation.
# (h.CV <- compak_CVbandwidth(days))
compak_evalpmf(days, 20:40, h = 0.0250636, parallel = FALSE)
}
