% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compak_fitpmf.R
\name{compak_fitpmf}
\alias{compak_fitpmf}
\title{Fit compak smoother at values x using a sample of data}
\usage{
compak_fitpmf(x, a.sample, h = NULL, bandwidth = c("KL", "CV"), ...)
}
\arguments{
\item{x}{numeric vector: the points of the grid at which the density is to be estimated.}

\item{a.sample}{numeric vector: the data sample from which estimate is to be computed.}

\item{h}{numeric: the bandwidth or smoothing parameter.}

\item{...}{other arguments passed to bandwidth selection, such as \code{from} and \code{to}.}
}
\value{

}
\description{
The function estimates the p.m.f. using the Conway-Maxwell-Poisson associated kernels (compak) smoother. Bandwidth can be specified as \code{h} or one can be provided by using cross-validation or minimising Kullback-Leibler divergence.
}
\examples{

}
