% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compak_graphics.R
\name{compak_graphics}
\alias{compak_graphics}
\alias{autoplot.compak}
\alias{plot.compak}
\title{Plot for an \code{compak} object in ggplot}
\usage{
\method{autoplot}{compak}(x, ...)

\method{plot}{compak}(x, ...)
}
\arguments{
\item{x}{an object class 'compak' object, obtain from a call to \code{compak_fitpmf()}.}

\item{...}{other argument passed to or from other methods (currently unused).}
}
\value{
An ggplot object for autoplot.
}
\description{
This function provides a histogram corresponding to the observations and then overlaid with the
pmf of a compak smoother using ggplot or graphics.
}
\examples{
data(days)
fit <- compak_fitpmf(days, bandwidth = "CV")
plot(fit) # or autoplot(fit)
}
