% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compak_KL.R
\name{compak_KL}
\alias{compak_KL}
\title{Calculate the Kullback-Leibler divergence between two density functions}
\usage{
compak_KL(f1, f2)
}
\arguments{
\item{f1, f2}{numeric vectors: two density functions}
}
\value{
The Kullback-Leibler distance between the two density functions.
}
\description{
Calculate the Kullback-Leibler divergence between two density functions
}
\keyword{internal}
