% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compak_KLbandwidth.R
\name{compak_KLbandwidth}
\alias{compak_KLbandwidth}
\title{Calculates the optimal bandwidth via minimising Kullback-Leibler divergence}
\usage{
compak_KLbandwidth(a.sample, from = 0, to = 1, parallel = FALSE)
}
\arguments{
\item{a.sample}{numeric vector; the data sample from which the estimate is to be computed.}

\item{from, to}{numeric; the range over which to optimize the bandwidth.}

\item{parallel}{logical: if \code{FALSE} (defualt), parallel process will not be used to evaluate the compak smoother.}
}
\value{
The optimal bandwidth parameter within the range.
}
\description{
Calculates the optimal bandwidth via minimising Kullback-Leibler divergence
}
\examples{
days <- c(rep(25,5), rep(26,5),rep(27,7),rep(28,8),
rep(29,11),rep(30,2),rep(31,1),rep(32,4),
rep(33,4),rep(34,2),rep(35,2))
h.KL <- compak_KLbandwidth(days, from = 0.025, parallel=F)

}
