% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compak_lambdaZ.R
\name{compak_lambdaZ}
\alias{compak_lambdaZ}
\title{Retrieve logZ and logLambda values from the precomputed data frame}
\usage{
compak_lambdaZ(mu, nu, logZ, logLambda)
}
\arguments{
\item{mu}{numeric: a vector of integer.}

\item{nu}{numeric: a vector of integer.}

\item{logZ}{numeric: a matrix of precomputed values for logZ.}

\item{logLambda}{numeric: a matrix of precomputed values for logLambda.}
}
\value{
A list containing the following components:

\item{logLambda}{a numeric vector of log-lambda values based on mu and nu.}
\item{logZ}{a numeric vector of log-Z values based on mu and nu.}
}
\description{
Retrieve logZ and logLambda values from the precomputed data frame
}
\keyword{internal}
