% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compak-package.R
\docType{data}
\name{days}
\alias{days}
\title{Days data set}
\format{
A vector of 51 observations.
}
\usage{
data(days)
}
\description{
This data set gives the count of days required for the insect pest (spirally whitefly, Aleurodicus dispersus Russel) to develop from egg to adult stages while being hosted on the Hura fruit tree plant (Hura crepitans).The \code{days} data has 51 observations.
}
\examples{
## For examples see example(compak_fitpmf)
}
\references{
Kiessé, T.S. (2017). On finite sample properties of nonparametric discrete asymmetric kernel estimators. \emph{Statistics}, 51, 1046--1060.
}
\keyword{datasets}
