% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compak_graphics.R
\name{gg_plot}
\alias{gg_plot}
\title{Plot for an \code{compak} object in ggplot}
\usage{
gg_plot(x, ...)
}
\arguments{
\item{x}{an object class 'compak' object, obtain from a call to \code{compak_fitpmf()}.}

\item{...}{other argument passed to or from other methods (currently unused).}
}
\value{
An ggplot object of the plot.
}
\description{
This function provides a histogram corresponding to the observations and then overlaid with the
pmf of a compak smoother using ggplot.
}
\examples{
data(days)
fit <- compak_fitpmf(days, bandwidth = "CV")
gg_plot(fit)
}
