% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compak_print.R
\name{print.compak}
\alias{print.compak}
\title{Print Values of COM-Poisson Model}
\usage{
\method{print}{compak}(x, digits = max(3, getOption("digits") - 3))
}
\arguments{
\item{x}{an object class 'compak', obtained from a call to \code{compak_fitpmf}.}

\item{digits}{numeric; minimum number of significant digits to be used for most numbers.}
}
\description{
\code{print} method for class \code{compak}.
}
\details{
\code{print.compak} can be used to print a short summary of object class 'compak'.
}
\examples{
## For examples see example(compak_fitpmf)

}
