% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compak-package.R
\docType{data}
\name{somites}
\alias{somites}
\title{somites data set}
\format{
A vector of 487 observations.
}
\usage{
data(somites)
}
\description{
This data set gives the number of body segments, known as somites, in common garden earthworms.
The \code{somites} data has 487 observations.
}
\examples{
## For examples see example(compak_fitpmf)
}
\references{
Owen, A.B. (2001). \emph{Empirical likelihood, Monographs on Statistics and Applied Probability (Series)}. Chapman & Hall, NY.

Pearl, R., & Fuller, W.N. (1905). Variation and correlation in the earthworm. \emph{Biometrika}, 4, 213–229.
}
\keyword{datasets}
