% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compak_CVbandwidth.R
\name{compak_CVbandwidth}
\alias{compak_CVbandwidth}
\title{Calculates the optimal bandwidth via cross-validation}
\usage{
compak_CVbandwidth(a.sample, interval = c(0.025, 1), parallel = F)
}
\arguments{
\item{a.sample}{numeric vector; the data sample from which the estimate is to be computed.}

\item{interval}{numeric vector; the end-points of the interval to be searched for the best bandwidth.}

\item{parallel}{logical: if \code{FALSE} (default), parallel process will not be used to evaluate the compak smoother.}
}
\value{
The optimal bandwidth parameter within the \code{interval}.
}
\description{
Calculates the optimal bandwidth via cross-validation
}
\examples{
data(days)
h.CV <- compak_CVbandwidth(days)
}
