% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compak_lambdaZ.R
\name{compak_lambdaZ}
\alias{compak_lambdaZ}
\title{Retrieve logZ and logLambda values from the precomputed data frame}
\usage{
compak_lambdaZ(mu, nu)
}
\arguments{
\item{mu}{numeric: a vector of integers that are greater than 1, representing the means}

\item{nu}{numeric: a vector of positive numbers, representing the dispersion}
}
\value{
A list containing the following components:

\item{logLambda}{a numeric vector of log-lambda values based on mu and nu.}
\item{logZ}{a numeric vector of log-Z values based on mu and nu.}
}
\description{
\code{logZ} and \code{logLambda} values are precomputed and stored in the \code{pc_data} data frame. This function would retrieve it based on the \code{mu} and \code{nu}.

The values are computed based on integer-value of \code{mu}>=1 and at a 0.1 grid of \code{nu}.
}
\keyword{internal}
